_05_paper_trading/requirements.txt
betfairlightweight>=2.20.0
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=10.0.0
xgboost>=1.7.0
lightgbm>=3.3.0
scikit-learn>=1.1.0
python-dotenv>=1.0.0
